+++ include/aikido/perception/PoseEstimatorModule.hpp
@@ -7,8 +7,8 @@
 #include <tf/transform_listener.h>
 #include "aikido/io/CatkinResourceRetriever.hpp"
 #include "aikido/io/yaml.hpp"
+#include "aikido/perception/AssetDatabase.hpp"
 #include "aikido/perception/DetectedObject.hpp"
-#include "aikido/perception/ObjectDatabase.hpp"
 #include "aikido/perception/PerceptionModule.hpp"
 
 namespace aikido {
@@ -19,7 +19,7 @@ namespace perception {
 /// It receives input from the PoseEstimator running in a separate process,
 /// by subscribing to a \c visualization_msg::MarkerArray ROS topic
 /// published by the PoseEstimator node.
-/// It uses a \c ObjectDatabase to resolve each marker to a
+/// It uses a \c AssetDatabase to resolve each marker to a
 /// \c dart::common::Uri, and updates the environment which is an
 /// \c aikido::planner::World.
 class PoseEstimatorModule : public PerceptionModule
@@ -33,7 +33,8 @@ public:
   ///	\param[in] nodeHandle The node handle to be passed to the detector
   ///	\param[in] markerTopic The name of the topic on which objects' pose
   /// information is being published
-  /// \param[in] configData The pointer to the loader of configuration data
+  /// \param[in] assetData The pointer to the loader of visual asset data (\see
+  /// AssetDatabase)
   /// \param[in] resourceRetriever A CatkinResourceRetriever for resources
   /// related to config files and models
   ///	\param[in] referenceFrameId The desired reference frame for the
@@ -43,19 +44,24 @@ public:
   PoseEstimatorModule(
       ros::NodeHandle nodeHandle,
       const std::string& markerTopic,
-      std::shared_ptr<ObjectDatabase> configData,
+      std::shared_ptr<AssetDatabase> assetData,
       std::shared_ptr<aikido::io::CatkinResourceRetriever> resourceRetriever,
       const std::string& referenceFrameId,
       dart::dynamics::Frame* referenceLink);
 
   virtual ~PoseEstimatorModule() = default;
 
-  // Documentation inherited
+  /// \copydoc PerceptionModule::detectObjects()
+  /// Looks for the following information sent via ROS (see \c DetectedObject):
+  /// visualization_msgs/Marker message like the following:
+  /// Marker.header.frame_id -> detectionFrameID
+  /// Marker.ns + "_" + Marker.id -> objUid (identity in DART world)
+  /// Marker.ns -> objAssetDBKey (determines visual asset, see AssetDatabase)
   bool detectObjects(
       const aikido::planner::WorldPtr& env,
-      std::vector<DetectedObject>& detectedObjects,
-      ros::Duration timeout = ros::Duration(0.0),
-      ros::Time timestamp = ros::Time(0.0)) override;
+      ros::Duration timeout = ros::Duration(),
+      ros::Time timestamp = ros::Time(0.0),
+      std::vector<DetectedObject>* detectedObjects = nullptr) override;
 
 private:
   /// For the ROS node that will work with the April Tags module
@@ -64,8 +70,8 @@ private:
   /// The name of the ROS topic to read marker info from
   std::string mMarkerTopic;
 
-  /// The pointer to the loader of configuration data
-  std::shared_ptr<ObjectDatabase> mConfigData;
+  /// The pointer to the loader of visual asset data
+  std::shared_ptr<AssetDatabase> mAssetData;
 
   /// To retrieve resources from disk and from packages
   std::shared_ptr<aikido::io::CatkinResourceRetriever> mResourceRetriever;