+++ src/perception/PoseEstimatorModule.cpp
@@ -14,13 +14,13 @@ namespace perception {
 PoseEstimatorModule::PoseEstimatorModule(
     ros::NodeHandle nodeHandle,
     const std::string& markerTopic,
-    std::shared_ptr<ObjectDatabase> configData,
+    std::shared_ptr<AssetDatabase> assetData,
     std::shared_ptr<aikido::io::CatkinResourceRetriever> resourceRetriever,
     const std::string& referenceFrameId,
     dart::dynamics::Frame* referenceLink)
   : mNodeHandle(std::move(nodeHandle))
   , mMarkerTopic(std::move(markerTopic))
-  , mConfigData(std::move(configData))
+  , mAssetData(std::move(assetData))
   , mResourceRetriever(std::move(resourceRetriever))
   , mReferenceFrameId(std::move(referenceFrameId))
   , mReferenceLink(std::move(referenceLink))
@@ -32,9 +32,9 @@ PoseEstimatorModule::PoseEstimatorModule(
 //=============================================================================
 bool PoseEstimatorModule::detectObjects(
     const aikido::planner::WorldPtr& env,
-    std::vector<DetectedObject>& detectedObjects,
     ros::Duration timeout,
-    ros::Time timestamp)
+    ros::Time timestamp,
+    std::vector<DetectedObject>* detectedObjects)
 {
   // Checks detected objects, looks up the database,
   // and adds new skeletons to the world env
@@ -62,6 +62,17 @@ bool PoseEstimatorModule::detectObjects(
 
   for (const auto& marker_transform : marker_message->markers)
   {
+    // TODO: Add DELETE_ALL Functionality
+    // TODO: Update when we move over to ROS Kinetic. Indigo doesn't
+    // have the enum value.
+    // if (marker_transform.action == visualization_msgs::Marker::DELETEALL) {
+    if (marker_transform.action == 3)
+    {
+      dtwarn << "[PoseEstimatorModule::detectObjects] We cannot currently "
+                "handle DELETE_ALL markers."
+             << std::endl;
+      continue;
+    }
     const auto& marker_stamp = marker_transform.header.stamp;
     const auto& obj_id = marker_transform.id;
     const auto& obj_ns = marker_transform.ns;
@@ -71,19 +82,40 @@ bool PoseEstimatorModule::detectObjects(
       continue;
     }
 
-    const std::string obj_uid = obj_ns + std::to_string(obj_id);
+    const std::string obj_uid = obj_ns + "_" + std::to_string(obj_id);
 
     // Initialize a DetectedObject class for this object
-    // and puts it into the output vector
-    DetectedObject this_object
-        = DetectedObject(obj_uid, detection_frame, marker_transform.text);
-    detectedObjects.push_back(this_object);
+    DetectedObject this_object(
+        obj_uid, obj_ns, detection_frame, marker_transform.text);
 
-    const std::string obj_db_key = this_object.getObjDBKey();
+    const std::string asset_key = this_object.getAssetKey();
+
+    if (asset_key.empty())
+    {
+      dtwarn << "[PoseEstimatorModule::detectObjects] Invalid YAML String in "
+                "Marker: "
+             << obj_uid << std::endl;
+      continue;
+    }
+
+    std::string obj_name;
+    dart::common::Uri obj_resource;
+    Eigen::Isometry3d obj_offset;
+    ros::Time t0 = ros::Time(0);
+
+    // Get the object name, resource, and offset from database by assetKey
+    try
+    {
+      mAssetData->getAssetByKey(asset_key, obj_name, obj_resource, obj_offset);
+    }
+    catch (std::runtime_error& e)
+    {
+      dtwarn << e.what() << std::endl;
+      continue;
+    }
 
     // If marker_transform.action is "DELETE",
     // remove a skeleton with obj_uid from env
-    // and move to next marker
     dart::dynamics::SkeletonPtr env_skeleton = env->getSkeleton(obj_uid);
     if (marker_transform.action == visualization_msgs::Marker::DELETE)
     {
@@ -93,14 +125,6 @@ bool PoseEstimatorModule::detectObjects(
       }
       continue;
     }
-
-    std::string obj_name;
-    dart::common::Uri obj_resource;
-    Eigen::Isometry3d obj_offset;
-    ros::Time t0 = ros::Time(0);
-
-    // get the object name, resource, and offset from database by objectKey
-    mConfigData->getObjectByKey(obj_db_key, obj_name, obj_resource, obj_offset);
 
     tf::StampedTransform transform;
     try
@@ -137,7 +161,6 @@ bool PoseEstimatorModule::detectObjects(
     // Check if skel in World
     // If there is, update its pose
     // If not, add skeleton to env
-    // A pose estimator module should provide the unique obj_uid per object
     if (env_skeleton == nullptr)
     {
       is_new_obj = true;
@@ -190,6 +213,11 @@ bool PoseEstimatorModule::detectObjects(
       env->addSkeleton(obj_skeleton);
     }
 
+    // Add object to output vector, if available
+    if (detectedObjects)
+    {
+      detectedObjects->push_back(this_object);
+    }
     any_detected = true;
   }